[FEAT/#13] 프로토콜 분리
protocol SearchResultsDelegate
protocol ViewModelType
파일 분리 완료
