+++ Movielity/Global/Protocols/ReusableProtocol.swift
@@ -17,15 +17,6 @@ extension UIView: ReusableProtocol {
     }
 }
 
-protocol ViewModelType {
-    associatedtype Input
-    associatedtype Output
-    
-    func transform(input: Input) -> Output
-}
 
-protocol SearchResultsViewControllerDelegate: AnyObject {
-    func searchResultsViewControllerDidSelectMovie(_ movie: IntoDetailMovieModel)
-}
